Additional fixes for MinGW 64bit.
- A few more int->intptr_t updates.
- Use winsock2.h to ensure sockets are the correct size for both 32bit and 64bit windows.
- SOCKET is not defined on *nix, so add conditional to typedef it to int.

+++ src/compat/regex.c
@@ -3747,7 +3747,7 @@ re_match_2 (bufp, string1, size1, string2, size2, pos, regs, stop)
 			  regstart[r] = old_regstart[r];
 
 			  /* xx why this test?  */
-			  if ((int) old_regend[r] >= (int) regstart[r])
+			  if ((intptr_t) old_regend[r] >= (intptr_t) regstart[r])
 			    regend[r] = old_regend[r];
 			}
 		    }
